+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.14.5
 
+- @iov/cli: Automatically import `TransactionId` from @iov/bcp.
 - @iov/crypto: Add `EnglishMnemonic.toString`; deprecate `.asString`.
 
 ## 0.14.4